fix: Remove circular dependency which caused test failures (BL-13906)
