+++ src/l10n/_locales/en/messages.json
@@ -50,9 +50,5 @@
     "Button.IgnoreImageDescriptions": {
         "message": "Ignore Image Descriptions",
         "description": "tooltip for the button to disable reading image descriptions"
-    },
-    "Video.BadVideoMessage": {
-        "message": "Sorry, this video cannot be played in this browser.",
-        "description": "message displayed when a video is invalid or missing"
     }
 }
